+++ booby/mixins.py
@@ -14,36 +14,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""The `errors` module contains all exceptions used by Booby."""
 
-
-class BoobyError(Exception):
-    """Base class for all Booby exceptions."""
-
-    pass
-
-
-class FieldError(BoobyError):
-    """This exception is used as an equivalent to :class:`AttributeError`
-    for :mod:`fields`.
-
-    """
-
-    pass
-
-
-class ValidationError(BoobyError):
-    """This exception should be raised when a `value` doesn't validate.
-    See :mod:`validators`.
-
-    """
-
-    pass
-
-
-class EncodeError(BoobyError):
-    pass
-
-
-class DecodeError(BoobyError):
-    pass
+class Encoder(object):
+    def encode(self):
+        return {
+            field.options.get('name', name): field.encode(getattr(self, name))
+                for name, field in self._fields.items()
+        }